requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
tzdata>=2024.1
google-auth-oauthlib>=1.2.0
caldav>=1.3.0
vobject>=0.9.7
anthropic>=0.39.0
